Add the new scrapers to a separate package
