+++ webapp/request_webscrapers/request_webscraper.py
@@ -1,0 +1,7 @@
+from webapp.data_retriever_interface import DataRetrieverInterface
+
+
+class RequestWebscraper(DataRetrieverInterface):
+
+    def __init__(self):
+        pass
